Fix and refractor search functionality
